requirements.txt
fastapi
uvicorn
pdfplumber
faiss-cpu
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
requests
python-multipart
numpy
gunicorn